feat: move `getpeername` to trait
